refactor(step1): 패키지명 변경
