+++ src/main/java/calculator/view/Input.java
@@ -3,7 +3,7 @@ package calculator.view;
 import java.util.InputMismatchException;
 import java.util.Scanner;
 
-import static calculator.validate.InputValidator.*;
+import static calculator.validator.InputValidator.*;
 
 public class Input {
     private final static Scanner SCANNER = new Scanner(System.in);
