+++ src/test/java/calculator/validator/InputValidatorTest.java
@@ -1,10 +1,9 @@
-package calculator.validate;
+package calculator.validator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.InputMismatchException;
